Add exception handling to the middleware
The change allows to set an explicit array of errors to be captured by
the tracer as errors.
Notice that errors are not muted by the middleware, they are re-raised
afterwards. So for example it's safe to capture interrupts or related,
and still they will bubble up.
By default all the standard exceptions are captured.

When an exception occurs, as described in the OT specification the
following operations are performed:

* we mark the span as failed by setting error tag to true
* we log the error by setting event to error, and error.object to the
thrown exception.